Renamings and make some of the scripts work
