+++ darepy/plotting/integration.py
@@ -7,8 +7,8 @@ patterns, radial intensity profiles, and azimuthal anisotropy information.
 import os
 import numpy as np
 import matplotlib.pyplot as plt
-from darepy_sans.utils import create_analysis_folder # Used for getting analysis folder path
-import integration as integ # Used for make_file_name
+from darepy.utils import create_analysis_folder # Used for getting analysis folder path
+from darepy import integration as integ # Used for make_file_name
 from scipy import optimize
 
 def plot_integ_radial(config, result, ScanNr, Frame, img_2D, data_azimuth):
